refactor!: cleanup for KeyboardEvent key polyfill and keys usages
BREAKING CHANGE:
 `keycodes.ts` constants for deprecated `keycode` removed, `key` property should be used instead
