+++ src/polyfills/polyfills.es6.ts
@@ -1,2 +1,3 @@
 import './list/dom.node.isConnected';
+import './list/dom.keyboard.key';
 import './list/dom.element.toggleAttribute';
